branch: branches to upgrade-redux, reorganizes file tree, collapses containers and components
