+++ browser/components/SinglePuppy.js
