+++ browser/index.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import AllPuppiesContainer from './AllPuppiesContainer';
+import AllPuppies from './AllPuppies';
 import { Provider } from 'react-redux';
 import store from './store';
 
@@ -10,7 +10,7 @@ ReactDOM.render(
   <Provider store={store} >
     <div className="container flexbox-container">
       <div className="jumbotron">
-        <AllPuppiesContainer />
+        <AllPuppies />
       </div>
     </div>
   </Provider>,
