+++ browser/redux/action-creators.js
@@ -1,4 +1,8 @@
+/* ---------- ACTION TYPES ---------- */
+
 export const GET_PUPPIES = 'GET_PUPPIES';
+
+/* ---------- ACTION CREATORS ---------- */
 
 export const getPuppies = puppies => {
   return {
