+++ browser/redux/reducer.js
