+++ browser/store.js
@@ -1,4 +1,4 @@
-import {reducer} from './reducer';
+import { reducer } from './redux/reducer';
 import { createStore } from 'redux';
 
 export default createStore(reducer);
